hw2-oscar-shih/requirements.txt
certifi==2021.10.8
cudatookit==11.3.1
cycler==0.10.0
imageio==2.9.0
joblib==0.17.0
kiwisolver==1.3.1
matplotlib==3.4.3
numpy==1.21.2
pandas==1.1.3
Pillow==8.4.0
pyparsing==2.4.7
python==3.8
python-dateutil==2.8.1
pytorch==1.10.0
pytz==2020.1
scikit-learn==0.23.2
scikit-image==0.18.1
scipy==1.6.2
six==1.16.0
torchvision==0.11.1
